+++ src/app/viewer.ts
@@ -50,6 +50,7 @@ import { PDBeDomainAnnotations } from './domain-annotations/behavior';
 import * as Complexes from './extensions/complexes';
 import * as Foldseek from './extensions/foldseek';
 import * as Interactions from './extensions/interactions';
+import { toMVSX } from './extensions/mvs';
 import { StateGallery, StateGalleryExtensionFunctions } from './extensions/state-gallery/behavior';
 import { StateGalleryManager } from './extensions/state-gallery/manager';
 import { StateGalleryControls } from './extensions/state-gallery/ui';
@@ -1114,7 +1115,7 @@ export class PDBeMolstarPlugin {
         Complexes: Complexes,
         Foldseek: Foldseek,
         Interactions: Interactions,
-        MVS: { MVSData, loadMVS },
+        MVS: { MVSData: { ...MVSData, toMVSX }, loadMVS },
         StateGallery: StateGalleryExtensionFunctions,
     };