+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 All notable changes to this project will be documented in this file, following the suggestions of [Keep a CHANGELOG](http://keepachangelog.com/). This project adheres to [Semantic Versioning](http://semver.org/) for its most widely used - and defacto - public interfaces.
 
 ## [Unreleased]
+- Experimental: Added `MVSData.toMVSX`
 
 ## [v3.10.0] - 2026-01-23
 - Mol* core dependency updated to 5.6.1